+++ pages/register.vue
